+++ CHANGELOG.md
@@ -21,8 +21,8 @@ and this project adheres to [Semantic Versioning][semver].
   and properly unbox `ListProxy` objects.
   This changes the return signature to `Iterator[tuple[str, list[str]]]`
   ([#398])
-- Filter `get_packages_from_bom` to return only package instances from
-  BOM root elements and support multi-root BOMs ([#423])
+- Filter `get_packages_from_bom` and `get_boms_from_spdx_document` to return
+  matching instances (empty list on none) and support multi-root BOMs ([#423])
 
 [#398]: https://github.com/spdx/ntia-conformance-checker/pull/398
 [#423]: https://github.com/spdx/ntia-conformance-checker/pull/423